requirements.txt
Django==4.2
django-cors-headers
djangorestframework
drf-yasg
pillow
psycopg2-binary